[REFACTOR] token dto 네이밍 수정
