--- src/main/java/org/example/domain/member/dto/response/AuthResponseDto.java
@@ -1,17 +1,0 @@
-package org.example.domain.member.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-public class AuthResponseDto {
-
-    @Builder
-    @Getter
-    @AllArgsConstructor
-    public static class TokenInfo {
-        private String accessToken;
-        private String refreshToken;
-    }
-
-}
