+++ src/main/java/org/example/domain/member/dto/response/TokenInfo.java
@@ -1,0 +1,13 @@
+package org.example.domain.member.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class TokenInfo {
+    private String accessToken;
+    private String refreshToken;
+}
